- 消除仿真层所有 f32 泄漏（万分比整数）
- 黄金确定性测试（93 测试全部通过）
- Replay 录制/回放/播放器 UI
- Replay 快退/快进/倍速控制

+++ crates/simulation/src/city/config.rs
@@ -17,12 +17,12 @@ pub struct AuraConfig {
 pub struct CityGlobalConfig {
     pub level_hp_multiplier: u32,
     pub base_population_per_level: u32,
-    pub visual_radius_base: f32,
-    pub visual_radius_per_level: f32,
-    pub heal_ratio: f32,
-    pub level_up_cost_multiplier: f32,
-    pub level_up_gain_ratio: f32,
-    pub capture_hp_ratio: f32,
+    pub visual_radius_base: u32,
+    pub visual_radius_per_level: u32,
+    pub heal_ratio: u32,
+    pub level_up_cost_multiplier: u32,
+    pub level_up_gain_ratio: u32,
+    pub capture_hp_ratio: u32,
     pub aura: AuraConfig,
 }
