+++ crates/simulation/src/map/mod.rs
@@ -7,6 +7,7 @@ use crate::soldier::*;
 use crate::types::*;
 use bevy_ecs::prelude::Resource;
 use bevy_ecs::world::World;
+use serde::{Deserialize, Serialize};
 
 /// Boundary wall limits — units can't pass.
 #[derive(Clone, Copy, Debug, Resource)]
@@ -18,8 +19,9 @@ pub struct WallBounds {
 }
 
 /// Map size preset.
-#[derive(Clone, Copy, PartialEq, Eq, Debug, Hash)]
+#[derive(Clone, Copy, PartialEq, Eq, Debug, Hash, Serialize, Deserialize, Default)]
 pub enum MapSize {
+    #[default]
     Small,  // 2000x2000
     Medium, // 3500x3500
     Large,  // 5000x5000
@@ -152,9 +154,7 @@ pub fn generate_map(world: &mut World, map_size: MapSize) {
             rng_range(&mut rng, level * 2, level * 5)
         };
         let max_population = level * city_config.base_population_per_level + pop_extra;
-        let visual_radius = (city_config.visual_radius_base
-            + level as f32 * city_config.visual_radius_per_level)
-            as u32;
+        let visual_radius = city_config.visual_radius_base + level * city_config.visual_radius_per_level;
 
         let unit_id = {
             let mut id_gen = world.resource_mut::<IdGenerator>();
